Add runqueries keybinding for explore
